created the DatabaseConfiguration class and Spring Profiles to change database depending on passed active profile
