+++ build.gradle
@@ -79,6 +79,7 @@ dependencies {
 java {
     toolchain {
         languageVersion = JavaLanguageVersion.of(15)
+        vendor = JvmVendorSpec.ADOPTOPENJDK
     }
 }
 
@@ -142,7 +143,7 @@ jacocoTestCoverageVerification {
         rule {
             limit {
                 counter = 'LINE'
-                minimum = 0.85
+                minimum = 0.84
             }
         }
     }
