+++ src/main/resources/application-dev.yml
@@ -1,0 +1,5 @@
+server:
+  port: 6666
+
+invoicing-system:
+  database: memory
