+++ src/main/resources/application-prod.yml
@@ -1,0 +1,5 @@
+server:
+  port: 9999
+
+invoicing-system:
+  database: file
