+++ src/main/resources/application.yml
@@ -1,0 +1,5 @@
+server:
+  port: 7777
+
+invoicing-system:
+  database: file
